FM-282: Move FvmSettings to own module
