+++ fendermint/app/settings/src/lib.rs
@@ -6,16 +6,17 @@ use config::{Config, ConfigError, Environment, File};
 use fvm_shared::econ::TokenAmount;
 use ipc_sdk::subnet_id::SubnetID;
 use serde::Deserialize;
-use serde_with::serde_as;
 use std::path::{Path, PathBuf};
 use tendermint_rpc::Url;
 
 use fendermint_vm_encoding::{human_readable_delegate, human_readable_str};
 
 use self::eth::EthSettings;
+use self::fvm::FvmSettings;
 use self::resolver::ResolverSettings;
 
 pub mod eth;
+pub mod fvm;
 pub mod resolver;
 
 /// Marker to be used with the `#[serde_as(as = "IsHumanReadable")]` annotations.
@@ -81,30 +82,6 @@ pub struct DbSettings {
     ///
     /// This affects how long we can go back in state queries.
     pub state_hist_size: u64,
-}
-
-#[serde_as]
-#[derive(Debug, Deserialize, Clone)]
-pub struct FvmSettings {
-    /// Overestimation rate applied to gas estimations to ensure that the
-    /// message goes through
-    pub gas_overestimation_rate: f64,
-    /// Gas search step increase used to find the optimal gas limit.
-    /// It determines how fine-grained we want the gas estimation to be.
-    pub gas_search_step: f64,
-    /// Indicate whether transactions should be fully executed during the checks performed
-    /// when they are added to the mempool, or just the most basic ones are performed.
-    ///
-    /// Enabling this option is required to fully support "pending" queries in the Ethereum API,
-    /// otherwise only the nonces and balances are projected into a partial state.
-    pub exec_in_check: bool,
-
-    /// Gas fee used when broadcasting transactions.
-    #[serde_as(as = "IsHumanReadable")]
-    pub gas_fee_cap: TokenAmount,
-    /// Gas premium used when broadcasting transactions.
-    #[serde_as(as = "IsHumanReadable")]
-    pub gas_premium: TokenAmount,
 }
 
 #[derive(Debug, Deserialize, Clone)]
